requirements.txt
branca==0.4.2
certifi==2020.12.5
chardet==4.0.0
folium==0.12.1
idna==2.10
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.2
Pillow==8.2.0
plum-py==0.3.1
requests==2.25.1
urllib3==1.26.4
